requirements.txt
streamlit
torch
ftfy
regex
tqdm
numpy
Pillow
faiss-cpu
scikit-learn
zipfile
requests